Refactored to new name of normflow(s) package
