+++ examples/aldp.py
@@ -4,14 +4,14 @@ import os
 import torch
 import numpy as np
 
-import normflow as nf
+import normflows as nf
 import boltzgen as bg
 
 from time import time
 from fab.utils.training import load_config
 from fab.target_distributions.aldp import AldpBoltzmann
 from fab import FABModel
-from fab.wrappers.normflow import WrappedNormFlowModel
+from fab.wrappers.normflows import WrappedNormFlowModel
 from fab.sampling_methods.transition_operators import HamiltonianMonteCarlo, Metropolis
 from fab.utils.aldp import evaluate_aldp
 from fab.utils.aldp import filter_chirality
@@ -197,7 +197,7 @@ for i in range(n_layers):
 # Map input to periodic interval
 layers.append(nf.flows.PeriodicWrap(ind_circ, bound_circ))
 
-# NormFlow model
+# normflows model
 flow = nf.NormalizingFlow(base, layers)
 wrapped_flow = WrappedNormFlowModel(flow).to(device)
