+++ fab/wrappers/__init__.py
@@ -1,3 +1,3 @@
-from fab.wrappers.normflow import WrappedNormFlowModel
+from fab.wrappers.normflows import WrappedNormFlowModel
 from fab.wrappers.nflow import WrappedNFlowsModel
 from fab.wrappers.torch import WrappedTorchDist
