+++ fab/wrappers/normflow.py
@@ -1,12 +1,12 @@
 from typing import Tuple
 
 import torch
-from normflow import NormalizingFlow
+from normflows import NormalizingFlow
 
 from fab.trainable_distributions import TrainableDistribution
 
 class WrappedNormFlowModel(TrainableDistribution):
-    """Wraps the distribution from normflow library
+    """Wraps the distribution from normflows library
     (https://github.com/VincentStimper/normalizing-flows) to work in this fab library."""
 
     def __init__(self, normalising_flow: NormalizingFlow):
