+++ requirements.txt
@@ -1,6 +1,6 @@
 numpy
 torch
-normflow @ git+https://github.com/VincentStimper/normalizing-flows.git
+normflows
 boltzgen @ git+https://github.com/VincentStimper/boltzmann-generators.git
 mdtraj
 nflows
